Address review: move firewall/SSH into InstanceLifecycle, accept String cpu
Moves applyProxyOnlyFirewall and injectSshKeyIfAvailable from BranchCommand
into InstanceLifecycle, eliminating the backwards lifecycle→command dependency.
Changes applyResourceLimits to accept String cpu to preserve original behavior
(no Integer.parseInt that could throw on non-integer TUI input). Hardens
getUid() with exit code and numeric validation.
