+++ src/main/java/dev/incusspawn/command/BranchCommand.java
@@ -9,7 +9,6 @@ import dev.incusspawn.incus.ResourceLimits;
 import dev.incusspawn.lifecycle.InstanceLifecycle;
 import dev.incusspawn.lifecycle.InstanceType;
 import dev.incusspawn.proxy.CertificateAuthority;
-import dev.incusspawn.proxy.MitmProxy;
 import dev.incusspawn.proxy.ProxyHealthCheck;
 import jakarta.inject.Inject;
 import picocli.CommandLine.Command;
@@ -19,7 +18,6 @@ import picocli.CommandLine.Parameters;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.List;
 @Command(
         name = "branch",
         description = "Create a new instance from an existing one",
@@ -79,7 +77,7 @@ public class BranchCommand implements Runnable {
         System.out.println("Branching '" + name + "' from '" + resolvedSource + "'...");
         incus.copy(resolvedSource, name);
 
-        var cpu = cpuLimit != null ? cpuLimit : ResourceLimits.adaptiveCpuLimit();
+        var cpu = String.valueOf(cpuLimit != null ? cpuLimit : ResourceLimits.adaptiveCpuLimit());
         var memory = memoryLimit != null ? memoryLimit : ResourceLimits.adaptiveMemoryLimit();
         var disk = diskLimit != null ? diskLimit : ResourceLimits.defaultDiskLimit();
 
@@ -307,43 +305,6 @@ public class BranchCommand implements Runnable {
         return NetworkMode.FULL;
     }
 
-    /**
-     * Apply iptables rules inside the container to restrict outbound traffic to only
-     * the host MITM proxy and DNS. Called after the container is started.
-     */
-    public static void applyProxyOnlyFirewall(IncusClient incus, String name) {
-        var gatewayIp = incus.configGet(name, Metadata.PROXY_GATEWAY);
-        if (gatewayIp.isEmpty()) {
-            System.err.println("Warning: no proxy gateway configured, skipping firewall rules.");
-            return;
-        }
-
-        var mitmPort = MitmProxy.CONTAINER_FACING_PORT;
-        var healthPort = MitmProxy.DEFAULT_HEALTH_PORT;
-
-        System.out.println("Applying proxy-only firewall rules...");
-
-        // Allow loopback
-        incus.shellExec(name, "iptables", "-A", "OUTPUT", "-o", "lo", "-j", "ACCEPT");
-        // Allow established/related connections
-        incus.shellExec(name, "iptables", "-A", "OUTPUT", "-m", "conntrack",
-                "--ctstate", "ESTABLISHED,RELATED", "-j", "ACCEPT");
-        // Allow MITM TLS proxy (port 443)
-        incus.shellExec(name, "iptables", "-A", "OUTPUT", "-d", gatewayIp,
-                "-p", "tcp", "--dport", String.valueOf(mitmPort), "-j", "ACCEPT");
-        // Allow health check endpoint
-        incus.shellExec(name, "iptables", "-A", "OUTPUT", "-d", gatewayIp,
-                "-p", "tcp", "--dport", String.valueOf(healthPort), "-j", "ACCEPT");
-        // Allow DNS to gateway
-        incus.shellExec(name, "iptables", "-A", "OUTPUT", "-d", gatewayIp,
-                "-p", "udp", "--dport", "53", "-j", "ACCEPT");
-        // Drop everything else
-        incus.shellExec(name, "iptables", "-P", "OUTPUT", "DROP");
-
-        System.out.println("  Outbound traffic restricted to " + gatewayIp +
-                " ports " + mitmPort + " (MITM), " + healthPort + " (health), 53 (DNS)");
-    }
-
     private static String getUid() {
         try {
             var pb = new ProcessBuilder("id", "-u");
@@ -383,45 +344,4 @@ public class BranchCommand implements Runnable {
         System.err.println("Warning: instance " + container + " may not be fully ready.");
     }
 
-    public static void injectSshKeyIfAvailable(IncusClient incus, String name) {
-        var check = incus.shellExec(name, "test", "-f", "/home/agentuser/.ssh/authorized_keys");
-        if (!check.success()) return;
-
-        var home = System.getProperty("user.home");
-        Path pubKey = null;
-        for (var keyName : List.of("id_ed25519.pub", "id_ecdsa.pub", "id_rsa.pub")) {
-            var candidate = Path.of(home, ".ssh", keyName);
-            if (Files.exists(candidate)) {
-                pubKey = candidate;
-                break;
-            }
-        }
-        if (pubKey == null) {
-            System.out.println("  SSH is available but no public key found in ~/.ssh/");
-            System.out.println("  Add your key manually: ssh-copy-id agentuser@<container-ip>");
-            return;
-        }
-
-        try {
-            var keyContent = Files.readString(pubKey).strip();
-            var tmpKey = Files.createTempFile("isx-ssh-", ".pub");
-            try {
-                Files.writeString(tmpKey, keyContent + "\n");
-                incus.filePush(tmpKey.toString(), name, "/home/agentuser/.ssh/authorized_keys");
-                incus.shellExec(name, "chown", "agentuser:agentuser", "/home/agentuser/.ssh/authorized_keys");
-                incus.shellExec(name, "chmod", "600", "/home/agentuser/.ssh/authorized_keys");
-            } finally {
-                Files.deleteIfExists(tmpKey);
-            }
-        } catch (IOException e) {
-            System.err.println("  Warning: failed to inject SSH key: " + e.getMessage());
-            return;
-        }
-
-        var ipResult = incus.shellExec(name, "hostname", "-I");
-        if (ipResult.success()) {
-            var ip = ipResult.stdout().strip().split("\\s+")[0];
-            System.out.println("  SSH access: ssh agentuser@" + ip);
-        }
-    }
 }
